APIの修正と問題の削除APIの作成
